+++ core/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Added
 
+- [#612](https://github.com/embedded-graphics/embedded-graphics/pull/625) Added `ImagePixelGetter` trait.
 - [#625](https://github.com/embedded-graphics/embedded-graphics/pull/625) Added CSS colors and color conversions to `Rgb666` and `Bgr666`.
 - [#656](https://github.com/embedded-graphics/embedded-graphics/pull/656) Added `Rgb666` and `Bgr666` conversions.